FIX: Fix web crash when add new label in edit any
It use directives to use v-focus to focus on the input.
When user is editing any tag, the input of tag will call focus().
After clicking add new lable, there will be two input in the page,
which will call focus(), too. As a result, there will be an infinite
loop to call focus().

BFOP-235, BFOP-468