FIX: Remove Tooltip elem from DOM in unbind hook
Everytime we mount an element with v-tooltip plugin,
a tooltip is appended to DOM tree.
However, if we do not remove the tooltip, and use v-if to toggle the
element's display, duplicate tooltips will be appended.
Therefore, we use v-directive's unbind hook
When unbind hook is triggered, we remove tooltip from DOM tree